Auto-detect file format in graph mode when implicit
The behavior of the command-line format/output paramete was misleading
and led to users creating broken files (e.g., a file ending in ".png"
that contained dot-syntax). We can do better by determining the format
by auto-detection, but still have the possibility to do a user override
if they wish to do so.