Allow buildchain to generate PKCS#12 files
Add an option for easy generation of PKCS#12 files. This way, you only
need to specify the certificate and private key that you want to end up
in the certificate, and buildchain will take care to select the correct
CA chain for you and embed it into the PKCS#12 file.