Added experimental code that examines X.509 certificates
I'd like to have a piece of code take a look at a certificate and tell
the user what is wrong. This is a very preliminary (and therefore not
yet available on the command line) draft towards this. It currently
examines the used crypto, only for RSA. I want to have the dimensions
that are analyzed in terms of bits security, commonness in the wild and
also a general verdict. For example, while RSA with e = 7 isn't per-se
insecure, it certainly is uncommon. Also want to have everything that's
output to be able to be output in JSON format to be machine readable so
it can be used by other tools (that make it look pretty). Anyways,
work-in-progress.