Implemented more examination of RSA moduli
Implemented scanning for public moduli such as the Debian SSH key moduli
or the RSA challenge moduli. They're stored internally as JSON and are
hashed using MD5 (to preserve space; we don't expect collisions for
valid moduli and even if there was deliberate tampering, it'd only ever
show the modulus as *weaker* than it would actually be).