More analysis on ECC keys
Also analyze field element width of elements. In theory, it would be
possible for someone to choose a particular field element in order to
have speed performances at the cost of security. This might be able to
detect it. We should also add tests for the average number of set bits
in X/Y as well.