Added ability to skip some time-consuming RSA testing
Tests like Miller-Rabin and Pollards Rho on RSA moduli are quite
expensive (order of ~10 seconds on a modern computer i7-5930K @ 3.5 GHz
with a 4096 bit modulus). Therefore, we want to skip those tests
selectively. Added that option in the form of generic analysis options,
which might come in handy for all kinds of other analyses as well.