Different order on Travis-CI caused failure
Travis-CI apparently first tested the pubkey handler, then the x509cert
handler. This caused the pubkey to be found and added to the Intervals()
class, but of course the inner one does not contain the outer Interval.
This should never have happened, since the _DER_HANDLER is a list and we
only checked names against the set. No idea why this is happening,
therefore. Changed it around a bit to be more flexible (custom precedence
ordering) and see if that changes anything.