Bug fixed in ScrapeEngine
Think I finally nailed it down! We search inside the ScrapeEngine for
out patterns and create a dict { offset: something } that we further
process down the line. In my Python version, when a dict's keys are just
integers, as they are here, they're always sorted automatically. That's
not guaranteed, however, and Travis' Python version has them in
different order, causing the later pubkey to be found before the
certificate and causing two matches instead of the expected one. Simply
added "sorted()", this should fix it!