Refactor SubprocessExecutor
Cleaned up the SubprocessExecutor API significantly and made it more
usable. Merging of stdout/stderr is not possible anymore, but that
merging is also never possible to do safely. Instead, we always capture
separate stdout/stderr FDs and merge them if so requested after-the-fact
(i.e., full stdout, then full stderr -- not mixing the two).