Updated the "graph" tool to be more useful
Actually what I wanted is to be able to visualize different aspects of a
PKI. This patch goes in that direction, but it's not complete yet. It
can visualize expiration date and certain aspects of the PKI (general
certificate type, key type, signature type), but the "security"
visualization which is probably most interesting is still missing --
it's dependent on finalization of the "examine" facility.

Well, first gotta properly refactor Ed25519/Ed448 certificates and clean
up all the Cryptosystem/KeySpecification stuff.

For this particular patch, grabbed the "AdvancedColorPalette" from
pygpufractal and made some minor changes to it.