Implement EdDSA support
Support for Ed25519 and Ed448 added. For this, needed some beefing up of
the ECC facility, since by default EdDSA private keys do not contain the
public key portion and need to be calculated on-the-fly. Use an affine
computation for this instead of projective one. Not all curves will
work, since Tonelli-Shanks isn't implemented fully and computation of
quadratic residues only is implemented in F_p for p % 8 == 5 or
p % 4 == 3 (Legendre formulae). Refactored the whole key handling
significantly in the process.