Made some more changes to use EdDSA in CAs
For root CAs, EdDSA now works. However, I cannot sign intermediate CAs
because when I leave out "default_md = sha256" in the openssl.cnf,
OpenSSL fails because of the missing config key. When I specify it, it
fails because of the invalid digest (also tried sha512, which /should/
be correct for Ed25519). So no idea how to handle this at the moment. Is
there a special value for "use what is implied by the signing
algorithm"?