Add ability to recognize wrong RSA parameters field
Something that I've seen in the wild is a certificate with RSA public
key that omits the "params" field from the Sequence. The standard says
that this field must be present and it must always be Null. BoringSSL
explicitly checks for presence of this (and NULLness of the value) and
therefore rejects certificates that fail this check.