Do not run into an exception when the sig/hash alg is unknown
Previously, we'd abort with an exception when the encountered signature
or hash function was underterminable. Let's report this as unusual
instead and don't make a security determinatioDo not run into an
exception when the sig/hash alg is unknown

Previously, we'd abort with an exception when the encountered signature
or hash function was underterminable. Let's report this as unusual
instead and don't make a security determinationn