Trying to add bit-bias checking for ECDSA R/S values
Ran into a bit of a problem here: We do not know the length of the R/S
values unless we know the curve. However, we don't know the curve
without having the parent certificate. Maybe we should add the parent
certificate also as an option, this would allow us to do even more
checks (such as if the SKI/AKI match and so on).