Preliminary support for explicit decoding of ECC domain params
Decoding and, also important, proper re-encoding of explicit ECC domain
parameters is quite difficult to get right, which is why I avoided
implementing it in the past. This adds some support for decoding, but
none for re-encoding (which would trigger false positive error messages
if done only half-way -- we're currently entirely ignoring the seed for
example).