Trying to fix Travis-CI tests
Not sure at all why this is happening on Travis, but not locally.
Locally I never see an ImportStatement class passed to the
_parse_import() class method, but on Travis that clearly happens.
Probably something odd with order of dictionary keys or something? Let's
catch that case and just return the identity in the parsing function and
see if that fixes things.