Fix issue with recursive importing of JudgementCode
When the "_require" attribute is not purged after an import, it means
every importer needs to re-specify all the flags again. We don't want
that, at import time the flags are resolved and then requirements are
purged, meaning that at each level the requirement check is performed
(not just at the very end).