Found a nicer way to create "extensions" ASN.1 structure
I never liked that we had explicit tag numbers in x509sak code, this is
not the right place. Found a way to reference the object type from the
original pyasn1_modules spec. This makes nicer code and cleans up one
ugly part that has nagged me for a while.