Fix certificate signature checks of expired certs
Root certs which have expired would erroneously report that they are not
self-signed, causing many issues in the test suite. This is apparently
due to a change in behavior of OpenSSL verify. The "verify" facility in
more recent OpenSSL versions (1.1.1j on my system) offers an option to
ignore the timestamp entirely, which fixes this. We still need to update
the dependency on that newer OpenSSL version.