Refactoring of signature check
Move signature check into the OpenSSLTools class where we'll have the
option to disable the time verification for older versions of OpenSSL.
This is necessary because we now have Travis build failures (most likely
because of old OpenSSL which doesn't support the new "ignore time"
option). Also needed to move the "certificate from TLS server"
functionality to avoid a circular dependency.